Allow overriding CC for cross-compilation. (#262)
In the case of SDL2 cross-compilation, I had to make this change.
FTE_TARGET=SDL2_arm and such don't work. If I override the compiler with
something like:
make CC=aarch64-linux-gnu-gcc then this particular line reverts it to
gcc and calls the system compiler anyway. I'm not sure why this line
needs to exist at all.

I did the same for windres, because it might also make sense to override
it with an i686-w64-mingw32-windres. Or potentially for the
aarch64-w64-mingw32 gcc in future.